Starter på ny deltakerdetaljer side for tiltakskoordinator
